Move string transform functions to a separate util class [NO COMPILATION]
+ Simplifies SortString
